backend/requirements.txt
fastapi[standard]
uvicorn[standard]
psycopg2-binary
sqlalchemy==2.0.18
python-dotenv==0.21.0